% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboPlot.R
\name{mboPlot}
\alias{mboPlot}
\title{mboPlot}
\description{
This is the base class for plotting in this package.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{opt_state}}{(\code{OptState})\cr
Environment containing necessary information needed during optimization in MBO.}

\item{\code{param_set}}{(\link{ParamSet})\cr
Object describing the parameter space of the search.}

\item{\code{param_vals}}{(\link{Param})\cr
Object containing the parameter values.}

\item{\code{param_types}}{('char()')\cr
Object containing the parameter values.}

\item{\code{control}}{(\link{MBOControl})\cr
Control object for mbo.}

\item{\code{shiny_uis}}{('html') \cr
Shiny ui elements.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mboPlot$new()}}
\item \href{#method-set_param_vals}{\code{mboPlot$set_param_vals()}}
\item \href{#method-plot}{\code{mboPlot$plot()}}
\item \href{#method-generateParamUiShiny}{\code{mboPlot$generateParamUiShiny()}}
\item \href{#method-clone}{\code{mboPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Pins the parameters to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mboPlot$new(opt_state, param_set, param_vals, control)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState})}

\item{\code{param_set}}{(\link{ParamSet})}

\item{\code{param_vals}}{(\link{Param})}

\item{\code{control}}{(\link{MBOControl})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_param_vals"></a>}}
\subsection{Method \code{set_param_vals()}}{
Sets parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mboPlot$set_param_vals(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\link[=ParamSet]{Param})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Some plot function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mboPlot$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateParamUiShiny"></a>}}
\subsection{Method \code{generateParamUiShiny()}}{
Generates ui elements of all parameters in the parameter set. The html elements are attached
to the object which calls the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mboPlot$generateParamUiShiny()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
html of ui for each parameter of the param_set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mboPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
