% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotSearchSpace.R
\name{MboPlotSearchSpace}
\alias{MboPlotSearchSpace}
\title{MboPlotSearchSpace}
\description{
This class generates plots for the visualization of the input space given
prior and posterior distributions of the evaluated parameters in the mbo run.
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotSearchSpace}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboPlotSearchSpace$new()}}
\item \href{#method-plot}{\code{MboPlotSearchSpace$plot()}}
\item \href{#method-clone}{\code{MboPlotSearchSpace$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotSearchSpace$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Plots prior distributions of mbo run specified in the set of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotSearchSpace$plot(
  include_y = self$param_vals$include_y,
  include_init_design = self$param_vals$include_init_design,
  search_space_components = getParamIds(self$opt_state$opt.path$par.set)[1:2]
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_y}}{(\code{boolean(1)})
Specifies if points should be coloured with y values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotSearchSpace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
