% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboShiny.R
\name{MboShiny}
\alias{MboShiny}
\title{MboShiny}
\description{
This class transforms inputs into shiny uis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mbo_plot}}{(\link{MboPlot})\cr
Object of class \code{MboPlot}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboShiny$new()}}
\item \href{#method-generatePlotParamUi}{\code{MboShiny$generatePlotParamUi()}}
\item \href{#method-generateSummaryTable}{\code{MboShiny$generateSummaryTable()}}
\item \href{#method-clone}{\code{MboShiny$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboShiny$new(mbo_plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mbo_plot}}{(\link{MboPlot}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generatePlotParamUi"></a>}}
\subsection{Method \code{generatePlotParamUi()}}{
Generates ui elements of all parameters in the parameter set. The html elements are attached
to the object which calls the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboShiny$generatePlotParamUi()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{(\code{logical()})
If TRUE, the output will only be saved in the object and not be returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{html}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateSummaryTable"></a>}}
\subsection{Method \code{generateSummaryTable()}}{
Generates a table of the MboSummary. The html elements are attached to the object
which calls the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboShiny$generateSummaryTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{html}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboShiny$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
