% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlot.R
\name{MboPlot}
\alias{MboPlot}
\title{MboPlot}
\description{
This is the base class for plotting in this package.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{opt_state}}{(\link{OptState})\cr
Environment containing necessary information needed during optimization in MBO.}

\item{\code{opt_path}}{(\link{OptPath})\cr
Optimization path of the mbo run.}

\item{\code{param_set}}{(\link{ParamSet})\cr
Object describing the parameter space of the search.}

\item{\code{param_vals}}{(\link{Param})\cr
Object containing the parameter values.}

\item{\code{mbo_control}}{(\link{MBOControl})\cr
Control object for mbo.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboPlot$new()}}
\item \href{#method-set_param_vals}{\code{MboPlot$set_param_vals()}}
\item \href{#method-clone}{\code{MboPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlot$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_param_vals"></a>}}
\subsection{Method \code{set_param_vals()}}{
Sets parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlot$set_param_vals(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\link[=ParamSet]{Param}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
