% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotDistToNeighbor.R
\name{MboPlotDistToNeighbor}
\alias{MboPlotDistToNeighbor}
\title{MboPlotDistToNeighbor}
\description{
This class generates plots for the visualization of the explore-exploit trade-pff.
Therefore the distance of \code{f(x)}
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotDistToNeighbor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-plot}{\code{MboPlotDistToNeighbor$plot()}}
\item \href{#method-clone}{\code{MboPlotDistToNeighbor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="initialize">}\href{../../VisBayesOpt/html/MboPlot.html#method-initialize}{\code{VisBayesOpt::MboPlot$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Plots prior distributions of mbo run specified in the set of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotDistToNeighbor$plot(
  dist_measure = c("min", "max", "mean"),
  theme = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dist_measure}}{(\code{character(1)})
Defines the distance measure which is considered for the determination of the distance between the
difference between the values of \code{f(x)}. \code{min} takes the minimum of the \code{k} calculated distances to
the neighbors while \code{max} uses maximum distance and \code{mean} calculates arithmetic mean distance.}

\item{\code{theme}}{(\link{theme|gg})
A theme to specify the ggplot default.}

\item{\code{k}}{(\code{integer(1)})
Defines the number of neighbors conisdered for the difference calculation. For e.g. \code{k=1} only the
difference to the previous value is considered (i.e. lagged value) while \code{k>1} enables to consider
the distance to the next \code{k} neighbors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotDistToNeighbor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
