% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotDistToNeighbor.R
\name{MboPlotDistToNeighbor}
\alias{MboPlotDistToNeighbor}
\title{MboPlotDistToNeighbor}
\description{
This class generates plots for the visualization of the explore-exploit trade-pff.
Therefore the distance of \code{f(x)}
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotDistToNeighbor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboPlotDistToNeighbor$new()}}
\item \href{#method-plot}{\code{MboPlotDistToNeighbor$plot()}}
\item \href{#method-clone}{\code{MboPlotDistToNeighbor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotDistToNeighbor$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Plots prior distributions of mbo run specified in the set of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotDistToNeighbor$plot(
  dist_measure = self$param_vals$dist_measure,
  include_init_design = self$param_vals$include_init_design
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dist_measure}}{(\code{character(1) | "min"})\cr
Defines the distance measure which is considered for the determination of the distance between
the search space componentens. \code{min} takes the minimum of the \code{p} componentents, while \code{max} uses
maximum distance and \code{mean} calculates arithmetic mean distance. The default is \code{min}.
@param include_init_design (\code{logical(1) | TRUE})\cr
Decide whether to include to points from the initial design. If \code{TRUE} the initial design points
are separated by a vertical line. If \code{FALSE} they are not plotted. Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotDistToNeighbor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
