% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotEstimationUncertainty.R
\name{MboPlotEstimationUncertainty}
\alias{MboPlotEstimationUncertainty}
\title{MboPlotEstimationUncertainty}
\description{
This class plots the uncertainty for one model in the mbo run.
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotEstimationUncertainty}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboPlotEstimationUncertainty$new()}}
\item \href{#method-plot}{\code{MboPlotEstimationUncertainty$plot()}}
\item \href{#method-clone}{\code{MboPlotEstimationUncertainty$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotEstimationUncertainty$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Plots the estimation uncertainty (y_evaluated vs y_estimated) for all previous outcome values at an iteration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotEstimationUncertainty$plot(
  highlight_iter = self$param_vals$highlight_iter,
  predict_y_iter_surrogate = self$param_vals$predict_y_iter_surrogate
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{highlight_iter}}{(\code{integer(1) | 1})\cr
Specifies the iteration at which the uncertainty estimation is calculated.
#' @param predict_y_iter_surrogate (\code{boolean(1) | FALSE})\cr
Specifies if y_hat is predicted with the surrogate from the chosen iteration. If FALSE y_hat is taken from the optimization
path, i.e. predicted based on surrogate of the respective iteration. If TRUE we use the surrogate of 'highlight_iter' iteration
to predict all points based on the search space x again.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotEstimationUncertainty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
