% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotInputSpace.R
\name{MboPlotInputSpace}
\alias{MboPlotInputSpace}
\title{MboPlotInputSpace}
\description{
This class generates plots for the visualization of the input space given
prior and posterior distributions of the evaluated parameters in the mbo run.
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotInputSpace}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-plot}{\code{MboPlotInputSpace$plot()}}
\item \href{#method-clone}{\code{MboPlotInputSpace$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="initialize">}\href{../../VisBayesOpt/html/MboPlot.html#method-initialize}{\code{VisBayesOpt::MboPlot$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Plots prior distributions of mbo run specified in the set of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotInputSpace$plot(
  type = c("prior", "posterior", "overlay"),
  plot = c("distribution", "iteration"),
  theme = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(\code{character(1)})
Defines the information to be used for the plot and can be either \code{prior}, \code{posterior} or òverlay.
Prior uses a random desgin of the parameter set.\cr
Posterior uses the values the optimizer searched over during the mbo run.\cr
Overlay plots a combination of both, \code{prior} and \code{posterior} in a combined plot.}

\item{\code{plot}}{(\code{chatacter(1)})
Defines plot type. \code{distribution} plots the distributions of each variable in the design.
\code{iteration} plots the value of each variable in the design over the single iterations.}

\item{\code{theme}}{(\link{theme|gg})
A theme to specify the ggplot default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotInputSpace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
