% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotInputSpace.R
\name{MboPlotInputSpace}
\alias{MboPlotInputSpace}
\title{MboPlotInputSpace}
\description{
This class generates plots for the visualization of the input space given
prior and posterior distributions of the evaluated parameters in the mbo run.
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotInputSpace}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{opt_path}}{(\link{OptPath})\cr
Optimization path of the mbo run.}

\item{\code{param_set}}{(\link{ParamSet})\cr
Object describing the parameter space of the search.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboPlotInputSpace$new()}}
\item \href{#method-plotInputSpace}{\code{MboPlotInputSpace$plotInputSpace()}}
\item \href{#method-clone}{\code{MboPlotInputSpace$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotInputSpace$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotInputSpace"></a>}}
\subsection{Method \code{plotInputSpace()}}{
Plots prior distributions of mbo run specified in the set of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotInputSpace$plotInputSpace(
  type = c("prior", "posterior", "overlay"),
  theme = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(\code{character(1)})
Defines the information to be used for the plot and can be either \code{prior}, \code{posterior} or òverlay.
Prior uses a random desgin of the parameter set.\cr
Posterior uses the values the optimizer searched over during the mbo run.\cr
Overlay plots a combination of both, \code{prior} and \code{posterior} in a combined plot.}

\item{\code{theme}}{(\link{theme|gg})
A theme to specify the ggplot default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotInputSpace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
