% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboPlotOptPath.R
\name{MboPlotOptPath}
\alias{MboPlotOptPath}
\title{MboPlotOptPath}
\description{
This class plots the opt state for a specified iteration of the mbo run.
}
\section{Super class}{
\code{\link[VisBayesOpt:MboPlot]{VisBayesOpt::MboPlot}} -> \code{MboPlotOptPath}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stored_pdp}}{(\code{list()})\cr
List for storing already calculated search_space_component effect object for PDPs. Helps to avoid re-calculation in case
it was already done.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboPlotOptPath$new()}}
\item \href{#method-plot}{\code{MboPlotOptPath$plot()}}
\item \href{#method-clone}{\code{MboPlotOptPath$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VisBayesOpt" data-topic="MboPlot" data-id="set_param_vals">}\href{../../VisBayesOpt/html/MboPlot.html#method-set_param_vals}{\code{VisBayesOpt::MboPlot$set_param_vals()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotOptPath$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Plots the opt state for a specified iteration of the mbo run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotOptPath$plot(
  highlight_iter = self$param_vals$highlight_iter,
  search_space_component = self$param_vals$search_space_component,
  parallel = TRUE,
  se_factor = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{highlight_iter}}{(\code{integer(1) | NULL})\cr
Specifies the iteration to be highlighted. The default \code{NULL} does not highlight any iteration.}

\item{\code{search_space_component}}{(\code{character(1) | NULL})\cr
Specifies the search_space_component to be calculated in PDP for higher dimensions, i.e. number of features > 2.
The default \code{NULL} is only a valid input for the 1 dimensional case.}

\item{\code{parallel}}{(\code{boolean(1) | TRUE})\cr
Specifies if computation of PDP is done parallel, i.e. on multiple cores. The default states that multiple
cores are used.}

\item{\code{se_factor}}{(\code{numeric(1) | 1})\cr
Specifies the scaling factor for the uncertainty (standard error) estimate for the 1 dimensional case.
The default value is that no up/or downscaling is applied.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\link{ggplot}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboPlotOptPath$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
