% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboSummary.R
\name{MboSummary}
\alias{MboSummary}
\title{MboSummary}
\description{
This class generates a summary of all relevvant features of a mbo run
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{opt_state}}{(\code{OptState})\cr
Environment containing necessary information needed during optimization in MBO.}

\item{\code{mbo_control}}{(\link{MBOControl})\cr
Control object for mbo.}

\item{\code{mboSummary}}{(\code{list()})\cr
Summarizes specific parameters of mbo run.}

\item{\code{opt_path}}{(\link{OptPath})\cr
Optimization path of the mbo run.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboSummary$new()}}
\item \href{#method-getMboSummary}{\code{MboSummary$getMboSummary()}}
\item \href{#method-clone}{\code{MboSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initializes the parameters to of object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboSummary$new(opt_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opt_state}}{(\link{OptState}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMboSummary"></a>}}
\subsection{Method \code{getMboSummary()}}{
Summarizes all relevant characterisitcs of the mbo run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboSummary$getMboSummary(silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{(\code{logical()})
If TRUE, the output will only be saved in the object and not be returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{list()}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
